ci: explicit file list for goreleaser
This updates the _goreleaser_ config to specify the files that we want
in the archive (LICENSE and README + the binary). Prior to this, the
defaults were matching `readme*`, which unintentionally included our
source code files, since they match that file glob pattern.

Additionally, the goreleaser config file is formatted and the archive
`rlcp` option is set to address the upcoming deprecation:
https://goreleaser.com/deprecations/#archivesrlcp