fix: re-create deleted resources
This fixes the resources to create a new resource when the API responds
with a "NOTFOUND" message when reading a resource that exists in the
state - resources that were created by Terraform and then deleted
outside of Terraform.

Prior to this the provider would return an error that the resource
didn't exist, requiring a manual removal from the state.