feature: image upload
This adds a `readme_image` resource for uploading images to ReadMe.

This is not really stateful - the provider uploads and tracks the
response, but cannot update or delete an existing image. Updates to the
source image will upload a new image (with a new URL) and deletions will
only remove it from the Terraform state.

The images API is not part of the official and documented ReadMe API -
it's the endpoint used when uploading an image using the web interface
to edit a document.

Dependencies are updated (requires bump of readme-api-go-client).