fix: image path validation
The `ValidateConfig` method occurs early before interpolations. When
using certain values, such as `for_each`, as the source, the method was
failing to properly validate because the values were unknown.

This removes the `ValidateConfig` method. It's redundant anyway, since
the file is checked during the attribute validation and before creating
a resource.