feat: API spec data source filtering
Add a `filter` attribute to the `readme_api_specification` data source
that can be used to apply filtering to the query to retrieve an API
specification.

An `id` or `title` is still required, but `filter` can be used to set
additional criteria, particularly around the category an API
specification exists within.