feat: Sort api_specifications data source
Add a `sort_by` attribute to the `readme_api_specifications` data source
to optionally sort the results by `title` or `last_synced`. Sorting the
map by a key with Terraform is clunky. Doing this with the provider is a
convenience.

When `sort_by` is unspecified, the results are in the order they were
received by the API.