fix: Trim leading/trailing whitespace from docs
The ReadMe API removes leading and trailing whitespace from docs upon
upload. Prior to this commit, the provider would produce a warning about
"inconsistent results after applying" if `chomp()` and `trimspace()`
weren't used.

This adds a `body_clean` read-only attribute to the doc resource and
data source that represents the body string as synchronized with
readme. The existing `body` attribute cannot be changed, since it's a
configurable value.