fix: external doc slug changes
ReadMe allows a doc's slug to be modified in the web editor, but not in
the API. The API uses the slug as an argument to retrieve a doc, and the
provider used the slug from state to retrieve the doc. When the slug is
modified in the editor, the provider would consider the doc deleted and
attempt to re-create it.

This modifies the provider's behavior to fallback to looking for a doc
by ID, which performs a search of all docs and is handled by the API
client library. If a matching doc is found, the slug is updated in
state. If the doc isn't found by slug or by ID, it's considered
deleted and removed from state to trigger a re-creation.