fix: volatile 'user' attribute on docs
The 'user' attribute on doc and related resources is a read-only value
with the ID of the user that last modified the resource. This includes
users working in the web console and the API user used by Terraform.
There was a scenario that could cause an inconsistent state after apply
because this value changed between the initial read and the post-update
refresh. This updates the attribute to be marked for change whenever the
'updated_at' value changes (which is triggered when several other
criteria changes).