feat: ability to associate doc with slug
This provides a solution for managing the doc content for an API
specification by providing a new `use_slug` attribute on the doc
resource that can be set to the slug of an existing doc on ReadMe (one
that was created externally to Terraform, such as the automatic API spec
docs).

This is non-conventional because it's not fully stateful - we're
basically importing the existing doc by its slug and overwriting it,
creating a managed resource. Likewise, we can't safely delete it through
Terraform, since it's nested within the specification.