fix: doc attribute inconsistencies
Replace the `otherattributemodifier` package with a simpler `ModifyPlan`
receiver function for handling attribute changes when other attributes
change. Most doc attributes should cause the "revision", "updated_at",
"user", and "algolia" attributes to be unknown, since they are updated
when changes to a doc are published. This also applies to the 'version'
resource.

Prior to this, the body's value wasn't tracked properly due to
normalization by the ReadMe API. With this update, the `body` attribute
tracks the user-provided value while the `body_clean` attribute is the
normalized version that gets compared to the API.

An additional change adds a second post-apply read of the doc to ensure
the stable values from the API are used in the state.