fix: deleting nested docs
Fix deleting nested docs by querying the respective doc's category to find the
docs within it to identify doc relationships.

This is particularly useful when using the `use_slug` attribute to manage an
API specification doc, since this doc gets created automatically by ReadMe and
its management is split between a doc and an API specification resource.

To control the behavior, a `config` parameter is added to the **provider**
configuration to enable certain features. Adding this to the provider's
configuration instead of a resource parameter allows for it to be toggled
without requiring changes to the resource first. The default behavior is to
**not** delete child docs and instead throw an error. When set to true, child
docs are deleted and a warning is emitted.

The introduction of the `config` parameter requires a change across resources
to use this new structure. This also allows for more flexibility in the future
to add more configuration options.